requirements.txt
# ComfyUI Universal Image Loader Requirements
# These dependencies are typically already available in ComfyUI installations

# Core dependencies
torch>=1.9.0
numpy>=1.20.0
Pillow>=8.0.0

# Optional dependencies for enhanced functionality
# (usually already included with ComfyUI)
torchvision>=0.10.0

# Development dependencies (optional)
# black>=22.0.0
# pytest>=6.0.0
# pylint>=2.12.0
